抽象整理HttpBody类，fread模式下使用循环池提升内存性能
